requirements.txt
transformers
torch
streamlit
python-dotenv
requests
SpeechRecognition
pyttsx3